chore(server): unship vm and container runtime
