+++ packages/browseros-agent/apps/server/src/lib/agents/runtime/index.ts
@@ -21,17 +21,7 @@ export {
   getCodexRuntime,
   prepareCodexContext,
 } from './codex-host-process-runtime'
-export { ContainerAgentRuntime } from './container-agent-runtime'
 export { ActionNotSupportedError, RuntimeNotReadyError } from './errors'
-export {
-  type ConfigureHermesRuntimeOptions,
-  configureHermesRuntime,
-  ensureHermesRuntimeReady,
-  getHermesRuntime,
-  HermesContainerRuntime,
-  type HermesContainerRuntimeConfig,
-  prepareHermesContext,
-} from './hermes-container-runtime'
 export {
   HostProcessAgentRuntime,
   type HostProcessAgentRuntimeDeps,
