--- packages/browseros-agent/apps/server/src/lib/vm/index.ts
@@ -1,12 +1,0 @@
-/**
- * @license
- * Copyright 2025 BrowserOS
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-
-export * from './errors'
-export * from './lima-cli'
-export * from './lima-config'
-export * from './paths'
-export * from './telemetry'
-export * from './vm-runtime'
