+++ packages/browseros-agent/apps/server/tests/lib/agents/acpx-agent-adapter.test.ts
@@ -121,7 +121,7 @@ describe('prepareAcpxAgentContext', () => {
     )
   })
 
-  it('prepares Hermes with HERMES_HOME pointing at the in-container agent home (translated from the host path)', async () => {
+  it('prepares Hermes with HERMES_HOME pointing at the host agent home', async () => {
     const browserosDir = await mkdtemp(join(tmpdir(), 'browseros-adapters-'))
     tempDirs.push(browserosDir)
     const prepared = await prepareAcpxAgentContext({
@@ -134,11 +134,8 @@ describe('prepareAcpxAgentContext', () => {
       message: 'remember this',
     })
 
-    // HERMES_HOME must be the *container-side* path (under /data) so the
-    // hermes binary running inside the container can actually open it.
-    // The host-side seeded files are reachable via the bind mount.
     expect(prepared.commandEnv.HERMES_HOME).toBe(
-      '/data/agents/harness/hermes-agent/home',
+      join(browserosDir, 'agents', 'hermes', 'harness', 'hermes-agent', 'home'),
     )
     expect(prepared.commandEnv).not.toHaveProperty('AGENT_HOME')
     expect(prepared.commandEnv).not.toHaveProperty('CODEX_HOME')
