+++ packages/browseros-agent/apps/server/tests/lib/agents/hermes/hermes-paths-provider-map.test.ts
@@ -16,15 +16,15 @@ import {
 import { getHermesProviderMapping } from '../../../../src/lib/agents/hermes/hermes-provider-map'
 
 describe('Hermes adapter helpers', () => {
-  it('resolves Hermes state under the BrowserOS VM state root', () => {
+  it('resolves Hermes state under the BrowserOS agents directory', () => {
     const browserosDir = '/tmp/browseros-test'
 
     expect(getHermesHarnessHostDir(browserosDir)).toBe(
-      '/tmp/browseros-test/vm/hermes/harness',
+      '/tmp/browseros-test/agents/hermes/harness',
     )
     expect(
       getHermesAgentHomeHostDir({ browserosDir, agentId: 'agent-1' }),
-    ).toBe('/tmp/browseros-test/vm/hermes/harness/agent-1/home')
+    ).toBe('/tmp/browseros-test/agents/hermes/harness/agent-1/home')
   })
 
   it('writes per-agent provider config from the Hermes provider map', async () => {
